+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## current
 - some menu restructuring
 - added the ability to select all notes
+- the current notes folder isn't shown in the recent notes folders any more
 
 ## 0.40
 - you now are able to select multiple notes and remove them or copy/move them to a recent note folder